multi step signup

+++ src/App.js
@@ -10,9 +10,14 @@ import withFirebase from './firebase/withFirebase'
 import Dashboard from './components/layouts/Dashboard';
 import Product from './components/layouts/Product';
 
+import EmailUnverified from './pages/EmailUnverified';
+import FinishSignUp from './pages/FinishSignUp';
+
 // routes
 import {
   DASHBOARD,
+  EMAIL_UNVERIFIED,
+  FINISH_SIGNUP,
   PRODUCT,
 } from './constants/routes';
 
@@ -29,10 +34,20 @@ class App extends Component {
     this.listener = this.props.firebase.auth.onAuthStateChanged(currentUser => {
       console.log("auth state changed ✴️")
       if (currentUser) {
-        localStorage.setItem('currentUser', JSON.stringify(currentUser));
-        this.setState({
-          currentUser: currentUser
-        });
+        this.props.firebase.accountReference(currentUser.uid)
+        .get()
+        .then((doc) => {
+          if (doc.exists) {
+            currentUser = {...currentUser.toJSON(), ...doc.data()}
+            localStorage.setItem('currentUser', JSON.stringify(currentUser));
+            this.setState({
+              currentUser: currentUser
+            });
+          }
+        })
+        .catch((error) => {
+          console.log("rror:" + error)
+        })
       } else {
         localStorage.removeItem('currentUser');
         this.setState({
@@ -51,6 +66,8 @@ class App extends Component {
       <FirebaseAuthContext.Provider value={this.state.currentUser}>
         <BrowserRouter>
           <Switch>
+            <Route path={EMAIL_UNVERIFIED} component={EmailUnverified}/>
+            <Route path={FINISH_SIGNUP} component={FinishSignUp}/>
             <Route path={DASHBOARD} component={Dashboard}/>
             <Route path={PRODUCT} component={Product}/>
           </Switch>
